fajita: Update libfod extension function prototypes

+++ fod/FodExtension.cpp
@@ -20,6 +20,6 @@ uint32_t getFodZOrder(uint32_t, bool touched) {
     return touched ? 0xfc8 : 0xbd6;
 }
 
-uint32_t getFodUsageBits(uint32_t usageBits, bool) {
+uint64_t getFodUsageBits(uint64_t usageBits, bool) {
     return usageBits;
 }
